requirements.txt
psycopg2-binary==2.9.9
pymongo==4.8.0
python-dotenv==1.0.1
pandas==2.2.2
matplotlib==3.8.2
seaborn==0.13.0
jupyterlab==4.0.11
ipykernel==6.29.0
